Connecting list with view models

+++ app/src/main/java/com/juanocampo/mytaxy/test/model/domain/Taxi.kt
@@ -2,7 +2,7 @@ package com.juanocampo.mytaxy.test.model.domain
 
 import com.juanocampo.mytaxy.test.utils.delegate.model.RecyclerViewType
 
-data class Taxi(val id: Int, val lat: Long, val lng: Long): RecyclerViewType {
+data class Taxi(val id: Int, val lat: Float, val lng: Float): RecyclerViewType {
 
     override fun getDelegateId() = id
